minions配置移入react-core
