+++ packages/rxdrag/editor/react-core/package.json
@@ -44,6 +44,8 @@
     "@rxdrag/shared": "workspace:*",
     "@rxdrag/react-core": "workspace:*",
     "@rxdrag/react-fieldy": "workspace:*",
+    "@rxdrag/minions-schema": "workspace:*",
+    "@rxdrag/minions-runtime-react": "workspace:*",
     "lodash": "^4.17.21",
     "styled-components": "^5.3.9"
   }
