+++ packages/rxdrag/editor/react-core/src/contexts.ts
@@ -1,6 +1,9 @@
+import { ILocales } from "@rxdrag/locales";
 import { IDesignerEngine, IDocument, ITreeNode } from "@rxdrag/core";
-import { createContext } from "react";
+import { ReactNode, createContext } from "react";
 import { IReactComponents } from "@rxdrag/react-shared";
+import { ActivityMaterialCategory } from "@rxdrag/minions-schema";
+import { IControllerDefine } from "./types";
 
 export const DesignerEngineContext = createContext<IDesignerEngine | undefined>(undefined)
 //IFrame中的Component不能跟engine中一套，所以必须要单独处理
@@ -8,3 +11,13 @@ export const ComponentsContext = createContext<IReactComponents>({})
 export const DocumentContext = createContext<IDocument | undefined>(undefined)
 export const NodeContext = createContext<ITreeNode | undefined>(undefined)
 export const LockContext = createContext<boolean | undefined>(undefined)
+
+
+export interface IMinionOptions {
+  materials?: ActivityMaterialCategory<ReactNode>[],
+  locales?: ILocales,
+  propSetters?: IReactComponents,
+  controllers?: IControllerDefine[]
+}
+
+export const MinionOptionContext = createContext<IMinionOptions | undefined>(undefined)
