+++ packages/rxdrag/editor/react-core/src/hooks/index.ts
@@ -22,6 +22,7 @@ export * from "./useGetNode"
 export * from "./useLanguage"
 export * from "./useLocalesManager"
 export * from "./useLocked"
+export * from "./useMinionOptions"
 export * from "./useMonitor"
 export * from "./useNode"
 export * from "./useNodeChanged"
